refactor(dashboard-seller): move product containers view to seller module
