+++ apps/dashboard/src/modules/seller/components/ProductCard.vue
